+++ services/directory/src/Directory.API/Program.cs
@@ -2,6 +2,8 @@ using System.Text.Json.Serialization;
 using Directory.API.Middleware;
 using Directory.Application;
 using Directory.Infrastructure;
+using Directory.Infrastructure.Persistence;
+using Microsoft.EntityFrameworkCore;
 using Serilog;
 
 var builder = WebApplication.CreateBuilder(args);
@@ -23,6 +25,12 @@ builder.Services.AddApplication();
 builder.Services.AddInfrastructure(builder.Configuration);
 
 var app = builder.Build();
+
+using (var scope = app.Services.CreateScope())
+{
+    var db = scope.ServiceProvider.GetRequiredService<DirectoryDbContext>();
+    await db.Database.MigrateAsync();
+}
 
 if (app.Environment.IsDevelopment())
 {